pyproject.toml
[tool.poetry]
name = "prevent-poetry-downgrade"
version = "0.1.0"
description = ""
authors = ["ScDor <18174994+ScDor@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.8"
gitpython = "^3.1.43"
typer = "^0.12.3"
packaging = "^24.1"
poetry = "^1.8.3"


[tool.poetry.group.dev.dependencies]
pylint = "^3.2.5"

[tool.pylint]
disable = "C0301"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
